requirements.txt
debugpy==1.5.1
dill==0.3.6
dill==0.3.4
einops==0.4.1
faiss==1.7.2
faiss_gpu==1.7.2
GPUtil==1.4.0
GPUtil==1.4.0
h5py==3.8.0
humanize==4.6.0
hydra-core==1.3.2
numpy==1.23.4
omegaconf==2.3.0
pandas==1.5.1
Pillow==9.5.0
psutil==5.9.0
python-dotenv==1.0.0
PyYAML==6.0
Requests==2.30.0
scikit_image==0.19.2
scikit_learn==1.2.1
timm==0.8.3.dev0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
